requirements.txt
simple-salesforce 
fastapi 
uvicorn 
python-dotenv 
groq